requirements.txt
tinytuya==1.13.2
click==7.1.2
prometheus-client==0.9.0
attrs==20.3.0
